+++ packages/client/src/app/globals.css
@@ -54,14 +54,12 @@ body {
   border-radius: 30px;
   position: absolute;
   left: 8%;
-  top: 13%;
   box-shadow: inset 4px 4px 4px rgba(0, 0, 0, 0.5);
 }
 
 .minimap-container {
   position: absolute;
   left: 77%;
-  top: 13%;
   width: 202px;
   height: 248px;
   background: linear-gradient(
