--- packages/client/src/app/styles/title.css
@@ -1,115 +1,0 @@
-.Title {
-  position: absolute;
-  left: 27.45%;
-  right: 25.64%;
-  top: 6.6%;
-  bottom: 90.34%;
-}
-
-.left-line {
-  position: absolute;
-  left: 27.45%;
-  right: 68.47%;
-  top: 7.36%;
-  bottom: 91.26%;
-}
-
-.ll2 {
-  position: absolute;
-  left: 27.45%;
-  right: 68.47%;
-  top: 50%;
-  bottom: 91.26%;
-
-  width: 20px;
-
-  border: 1px solid #f9e05f;
-}
-
-.ll1 {
-  position: absolute;
-  left: 27.45%;
-  right: 68.47%;
-  top: 5%;
-  bottom: 92.64%;
-
-  width: 20px;
-
-  border: 1px solid #f9e05f;
-}
-
-.right-line {
-  position: absolute;
-  left: 70.28%;
-  right: 25.64%;
-  top: 7.36%;
-  bottom: 91.26%;
-}
-
-.rl2 {
-  position: absolute;
-  left: 70.28%;
-  right: 25.64%;
-  top: 50%;
-  bottom: 91.26%;
-  width: 20px;
-
-  border: 1px solid #f9e05f;
-}
-
-.rl1 {
-  position: absolute;
-  left: 70.28%;
-  right: 25.64%;
-  top: 5%;
-  bottom: 92.64%;
-  width: 20px;
-
-  border: 1px solid #f9e05f;
-}
-
-.title-text {
-  position: absolute;
-  left: 32.86%;
-  right: 31.05%;
-  top: 6.6%;
-  bottom: 90.34%;
-}
-
-.Whisper-In-A-Bottle1 {
-  position: absolute;
-  left: 32.86%;
-  right: 31.05%;
-  top: 6.6%;
-  bottom: 90.34%;
-
-  font-family: "Press Start 2P";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 20px;
-  white-space: nowrap;
-  /* identical to box height */
-
-  color: #f5f5f5;
-}
-
-.Whisper-In-A-Bottle2 {
-  position: absolute;
-  left: 32.86%;
-  right: 31.05%;
-  top: 6.6%;
-  bottom: 90.34%;
-
-  font-family: "Press Start 2P";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20px;
-  line-height: 20px;
-  white-space: nowrap;
-  /* identical to box height */
-
-  color: #f9e05f;
-
-  filter: blur(5px);
-}
