--- packages/client/src/app/styles/title.tsx
@@ -1,19 +1,0 @@
-import "./title.css";
-export const Title = () => {
-  return (
-    <div className="title w-6/5">
-      <div className="left-line">
-        <div className="ll1" />
-        <div className="ll2" />
-      </div>
-      <div className="title-text">
-        <h1 className="Whisper-In-A-Bottle1">Whisper In A Bottle</h1>
-        <h1 className="Whisper-In-A-Bottle2">Whisper In A Bottle</h1>
-      </div>
-      <div className="right-line">
-        <div className="rl1" />
-        <div className="rl2" />
-      </div>
-    </div>
-  );
-};
