+++ packages/client/src/components/background/cooling_fan.tsx
