+++ packages/client/src/components/background/copyright.tsx
