+++ packages/client/src/components/background/sound.tsx
