+++ packages/client/src/hooks/auth/use_login_hook.ts
@@ -8,7 +8,7 @@ import { REQUEST_PORT } from "@client/src/ts/enum/REQUEST_PORT";
 
 import { Player } from "@client/src/ts/interface/player.interface";
 
-import { ValidateId } from "../../../utils/validate_id";
+import { ValidateId } from "../../utils/validate_id";
 
 const LoginAxios = async (user: Player): Promise<boolean> => {
   try {
