+++ packages/client/src/styles/background/cooling_fan.css
