+++ packages/client/src/styles/background/copyright.css
